Modified repos to accept Integer as id instead of Long

+++ src/main/java/com/pet/server/repos/FoodRepository.java
@@ -3,4 +3,4 @@ package com.pet.server.repos;
 import com.pet.server.model.Food;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface FoodRepository extends JpaRepository<Food, Long> {}
+public interface FoodRepository extends JpaRepository<Food, Integer> {}
