+++ src/main/java/com/pet/server/repos/PetRepository.java
@@ -3,4 +3,4 @@ package com.pet.server.repos;
 import com.pet.server.model.Pet;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PetRepository extends JpaRepository<Pet, Long> {}
+public interface PetRepository extends JpaRepository<Pet, Integer> {}
